chore: fix typedoc after upgrade
Typedoc now expects root file(s) to be specified instead of a source
dir. We could still pass src/* but the resulting documentation for that
is quite messy and doesn't convey what the actual exported values from
module index are. Instead, specify pankkiyhteys.ts as the root (since
it's exported in index.js) and use typedoc-plugin-missing-exports to
include even the non-exported types in the documentation.